refactor config into folder
