+++ crates/cubecl-linalg/src/matmul/cmma/config/mod.rs
@@ -1,0 +1,4 @@
+mod base;
+mod strategy;
+
+pub(crate) use base::{CmmaConfig, ComptimeCmmaInfo};
